requirements.txt
django==5.1.4
djangorestframework==3.14
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.2.2
pandas==2.1.3
celery==5.3.1
redis==5.0.0
drf-spectacular==0.26.3
django-debug-toolbar==4.1
whitenoise==6.5.0
pillow==10.0.1
python-decouple==3.8
dj-rest-auth
django-allauth
reportlab
redis
celery
djoser
setuptools>=58.0.0
django-cors-headers>=3.6.0,<4.0
django-filter